refactor: changed components to app folder
