+++ src/app/apps/connections/ConnectionsPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Body } from '@/app/core/components/layout/Body';
+import { Body } from '@/app/components/layout/Body';
 import ConnectionTable from './components/ConnectionsTable';
 
 const ConnectionPage: React.FC = () => {
