+++ src/app/apps/contacts/ContactsPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Body } from '@/app/core/components/layout/Body';
+import { Body } from '@/app/components/layout/Body';
 import ContactTable from './components/ContactsTable';
 
 const ContactPage: React.FC = () => {
