+++ src/app/apps/messages/MessagesPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Body } from '@/app/core/components/layout/Body';
+import { Body } from '@/app/components/layout/Body';
 import MessageTable from './components/MessagesTable';
 
 const MessagePage: React.FC = () => {
